+++ tests/TestCase.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Nerrowake\Strata\Tests;
+
+use Nerrowake\Strata\StrataServiceProvider;
+use Orchestra\Testbench\TestCase as OrchestraTestCase;
+
+abstract class TestCase extends OrchestraTestCase
+{
+    protected function getPackageProviders($app): array
+    {
+        return [
+            StrataServiceProvider::class,
+        ];
+    }
+}